+++ sysbench/sb_rand.h
@@ -62,12 +62,12 @@ void sb_rand_done(void);
 void sb_rand_thread_init(void);
 
 /* Generator functions */
-int sb_rand_default(int, int);
-int sb_rand_uniform(int, int);
-int sb_rand_gaussian(int, int);
-int sb_rand_special(int, int);
-int sb_rand_pareto(int, int);
-int sb_rand_uniq(int a, int b);
+uint64_t sb_rand_default(uint64_t, uint64_t);
+uint64_t sb_rand_uniform(uint64_t, uint64_t);
+uint64_t sb_rand_gaussian(uint64_t, uint64_t);
+uint64_t sb_rand_special(uint64_t, uint64_t);
+uint64_t sb_rand_pareto(uint64_t, uint64_t);
+uint64_t sb_rand_uniq(uint64_t a, uint64_t b);
 void sb_rand_str(const char *, char *);
 
 #endif /* SB_RAND_H */
